Support for Oracle test database schema

+++ schemacrawler-testdb/src/main/resources/db/books/05_tables_04_A.sql
@@ -1,10 +1,10 @@
 -- Table with special characters in the name
 CREATE TABLE Βιβλία
 (
-  Μοναδικός BIGINT NOT NULL,
+  Μοναδικός SMALLINT NOT NULL,
   Τίτλος VARCHAR(255) NOT NULL,
   Περιγραφή VARCHAR(255),
-  Εκδότης BIGINT NOT NULL,
+  Εκδότης SMALLINT NOT NULL,
   CONSTRAINT PK_βιβλία PRIMARY KEY (Μοναδικός),  
   CONSTRAINT FK_βιβλία_Publishers FOREIGN KEY (Εκδότης) REFERENCES Publishers (Id)
 )
