+++ schemacrawler-testdb/src/main/resources/db/books/05_tables_04_B.sql
@@ -1,7 +1,7 @@
 -- Table with special characters in the name
 CREATE TABLE Βιβλία
 (
-  Μοναδικός BIGINT NOT NULL,
+  Μοναδικός SMALLINT NOT NULL,
   Τίτλος VARCHAR(255) NOT NULL,
   Περιγραφή VARCHAR(255),
   Εκδότης INTEGER NOT NULL,
